remove lr_find

+++ dataset/packing_dataset.py
@@ -45,7 +45,7 @@ class CombinedDatasetIterator:
         s = batch["vae_latent_shape"]
 
         vae_latent_shape = (s[0][0], s[1][0], s[2][0])
-        batch["vae_latent"] = torch.tensor(np.stack([np.frombuffer(s, dtype=np.float16).copy() for s in batch['vae_latent']])).reshape(-1, *vae_latent_shape),
+        batch["vae_latent"] = torch.tensor(np.stack([np.frombuffer(s, dtype=np.float32).copy() for s in batch['vae_latent']])).reshape(-1, *vae_latent_shape),
         batch["text_embedding"] = torch.tensor(np.stack([np.frombuffer(s, dtype=np.float16).copy() for s in batch['text_embedding']])),
 
         return batch
